+++ nmad/protocols/proto_mpi/src/mpi_nmad_private.c
@@ -45,9 +45,9 @@ static int 		     nb_incoming_msg    = 0;
 static int 		     nb_outgoing_msg    = 0;
 
 /** gives the outgoing gate attached to a node */
-static uint16_t             *out_gate_id	= NULL;
+static uint8_t             *out_gate_id	= NULL;
 /** gives the incoming gate attached to a node */
-static uint16_t             *in_gate_id	        = NULL;
+static uint8_t             *in_gate_id	        = NULL;
 /** gives the node attached to an outgoing gate */
 static int                  *out_dest	 	= NULL;
 /** gives the node attached to an incoming gate */
@@ -280,19 +280,19 @@ int mpir_internal_exit(void) {
   return MPI_SUCCESS;
 }
 
-uint16_t mpir_get_in_gate_id(int node) {
+uint8_t mpir_get_in_gate_id(int node) {
   return in_gate_id[node];
 }
 
-uint16_t mpir_get_out_gate_id(int node) {
+uint8_t mpir_get_out_gate_id(int node) {
   return out_gate_id[node];
 }
 
-int mpir_get_in_dest(uint16_t gate) {
+int mpir_get_in_dest(uint8_t gate) {
   return in_dest[gate];
 }
 
-int mpir_get_out_dest(uint16_t gate) {
+int mpir_get_out_dest(uint8_t gate) {
   return out_dest[gate];
 }
 
@@ -586,7 +586,7 @@ static inline int mpir_datatype_struct_split(void *recvbuffer,
  * Checks the current sending NM sequence and triggers flushing of
  * previous outgoing requests when needed.
  */
-static inline int mpir_check_send_seq(uint16_t gate_id,
+static inline int mpir_check_send_seq(uint8_t gate_id,
                                       uint8_t tag) {
   int seq = nm_so_sr_get_current_send_seq(p_so_sr_if, gate_id, tag);
   if (seq == NM_SO_PENDING_PACKS_WINDOW-1) {
@@ -604,7 +604,7 @@ static inline int mpir_check_send_seq(uint16_t gate_id,
  * Checks the current receiving NM sequence and triggers flushing of
  * previous incoming requests when needed.
  */
-static inline int mpir_check_recv_seq(uint16_t gate_id,
+static inline int mpir_check_recv_seq(uint8_t gate_id,
                                       uint8_t tag) {
   int seq = nm_so_sr_get_current_recv_seq(p_so_sr_if, gate_id, tag);
   int err;
@@ -811,7 +811,7 @@ int mpir_set_status(MPI_Request *request,
   if (mpir_request->request_type == MPI_REQUEST_RECV ||
       mpir_request->request_type == MPI_REQUEST_PACK_RECV) {
     if (mpir_request->request_source == MPI_ANY_SOURCE) {
-      uint16_t gate_id;
+      uint8_t gate_id;
       err = nm_so_sr_recv_source(p_so_sr_if, mpir_request->request_nmad, &gate_id);
       status->MPI_SOURCE = in_dest[gate_id];
     }
